First attempt at isolating transaction verification strategy
